+++ libs/store/src/lib/infrastructure/persistence/index.ts
@@ -1,3 +1,2 @@
 export * from './store/entity-store-server';
-export * from './entity-options.service'; 
-export * from './default-entity-options.service'; 
+export * from './utils';
