+++ libs/store/src/lib/infrastructure/persistence/store/index.ts
