+++ libs/store/src/lib/infrastructure/persistence/utils/index.ts
@@ -1,0 +1,1 @@
+export * from './entity-options.service';
